requirements.txt
streamlit
pandas
mysql-connector-python
altair
python-dotenv
bcrypt
psycopg2-binary